Added not null and unique constraints and checks, renamed sequences and changed some column types.

+++ src/main/java/com/kh013j/model/domain/Dish.java
@@ -6,6 +6,7 @@ import javax.persistence.*;
 import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.Size;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name = "dish", schema = "rh")
@@ -13,7 +14,6 @@ import javax.validation.constraints.Size;
 @ToString
 @AllArgsConstructor
 @NoArgsConstructor
-//@RestHubEntity(table = @Table(name="dish", schema = "rh"))
 public class Dish {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -27,7 +27,8 @@ public class Dish {
     @Min(0)
     private int preparingtime;
     @Max(100000)
-    private int price;
+    // change price type!
+    private BigDecimal price;
     @ManyToOne
     @JoinColumn(name = "category_id")
     private Category category;
